AUT-3530: Handle re-authenticating Users that are signed out or whose session has expired.
When a User starts a re-authentication journey they may have logged out from another RP or their OL session may have expired.  If that is the case we will not have access to a session in the user context and will not know which email address was originally used to sign in.

As we do not have access to their internal subject id we will use the rpPairwiseId sent in the re-authentication request to uniquely identify the User when updating authentication attempts.  This allows us to store a count should the user provide an incorrect email address. This could be an email not know to OL at all and an eamil address known to OL that, when used to calculate a rpPairwiseId, does not match the rpPairwiseId provided in the re-authentication request.
