+++ integration-tests/src/test/java/uk/gov/di/authentication/api/CheckReAuthUserHandlerIntegrationTest.java
@@ -14,22 +14,29 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import uk.gov.di.authentication.frontendapi.entity.CheckReauthUserRequest;
 import uk.gov.di.authentication.frontendapi.lambda.CheckReAuthUserHandler;
 import uk.gov.di.authentication.shared.entity.ClientSession;
+import uk.gov.di.authentication.shared.entity.CountType;
+import uk.gov.di.authentication.shared.entity.ErrorResponse;
+import uk.gov.di.authentication.shared.entity.JourneyType;
 import uk.gov.di.authentication.shared.entity.VectorOfTrust;
 import uk.gov.di.authentication.shared.helpers.ClientSubjectHelper;
+import uk.gov.di.authentication.shared.helpers.NowHelper;
 import uk.gov.di.authentication.shared.serialization.Json;
+import uk.gov.di.authentication.shared.services.AuthenticationAttemptsService;
 import uk.gov.di.authentication.sharedtest.basetest.ApiGatewayHandlerIntegrationTest;
 import uk.gov.di.authentication.sharedtest.extensions.AuthenticationAttemptsStoreExtension;
-import uk.gov.di.orchestration.shared.entity.ErrorResponse;
 import uk.gov.di.orchestration.sharedtest.helper.KeyPairHelper;
 
 import java.net.URI;
 import java.security.KeyPair;
+import java.time.Instant;
 import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 import java.util.Base64;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
+import java.util.stream.IntStream;
 
 import static java.util.Collections.singletonList;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -57,14 +64,34 @@ public class CheckReAuthUserHandlerIntegrationTest extends ApiGatewayHandlerInte
     public static final String ENCODED_DEVICE_INFORMATION =
             "R21vLmd3QilNKHJsaGkvTFxhZDZrKF44SStoLFsieG0oSUY3aEhWRVtOMFRNMVw1dyInKzB8OVV5N09hOi8kLmlLcWJjJGQiK1NPUEJPPHBrYWJHP358NDg2ZDVc";
 
+    private static final IntegrationTestConfigurationService CONFIGURATION_SERVICE =
+            new IntegrationTestConfigurationService(
+                    notificationsQueue,
+                    tokenSigner,
+                    docAppPrivateKeyJwtSigner,
+                    configurationParameters) {
+
+                @Override
+                public String getTxmaAuditQueueUrl() {
+                    return txmaAuditQueue.getQueueUrl();
+                }
+
+                @Override
+                public boolean isAuthenticationAttemptsServiceEnabled() {
+                    return true;
+                }
+            };
+
+    private final AuthenticationAttemptsService authenticationService =
+            new AuthenticationAttemptsService(CONFIGURATION_SERVICE);
+
     @BeforeEach
     void setup() throws Json.JsonException {
+
         var sessionId = redis.createAuthenticatedSessionWithEmail(TEST_EMAIL);
         requestHeaders = createHeaders(sessionId);
         redis.createClientSession(CLIENT_SESSION_ID, createClientSession());
-        handler =
-                new CheckReAuthUserHandler(
-                        TXMA_ENABLED_CONFIGURATION_SERVICE, redisConnectionService);
+        handler = new CheckReAuthUserHandler(CONFIGURATION_SERVICE, redisConnectionService);
         txmaAuditQueue.clear();
     }
 
@@ -90,7 +117,13 @@ public class CheckReAuthUserHandlerIntegrationTest extends ApiGatewayHandlerInte
 
     @Test
     void shouldReturn404WhenUserNotFound() {
-        var request = new CheckReauthUserRequest(TEST_EMAIL, "random-pairwise-id");
+        userStore.signUp(TEST_EMAIL, "password-1", SUBJECT);
+        registerClient("https://randomSectorIDuRI.COM");
+        byte[] salt = userStore.addSalt(TEST_EMAIL);
+        var expectedPairwiseId =
+                ClientSubjectHelper.calculatePairwiseIdentifier(
+                        SUBJECT.getValue(), INTERNAL_SECTOR_HOST, salt);
+        var request = new CheckReauthUserRequest(TEST_EMAIL, expectedPairwiseId);
         var response =
                 makeRequest(
                         Optional.of(request),
@@ -126,12 +159,14 @@ public class CheckReAuthUserHandlerIntegrationTest extends ApiGatewayHandlerInte
     void shouldReturn400WhenUserEnteredInvalidEmailTooManyTimes() {
         userStore.signUp(TEST_EMAIL, "password-1", SUBJECT);
         registerClient("https://randomSectorIDuRI.COM");
-
-        redis.incrementEmailCount(TEST_EMAIL);
-        redis.incrementEmailCount(TEST_EMAIL);
-        redis.incrementEmailCount(TEST_EMAIL);
-        redis.incrementEmailCount(TEST_EMAIL);
-        redis.incrementEmailCount(TEST_EMAIL);
+        int count = 5;
+        while (count-- > 0) {
+            authenticationService.createOrIncrementCount(
+                    SUBJECT.getValue(),
+                    NowHelper.nowPlus(10, ChronoUnit.MINUTES).toInstant().getEpochSecond(),
+                    JourneyType.REAUTHENTICATION,
+                    CountType.ENTER_EMAIL);
+        }
 
         byte[] salt = userStore.addSalt(TEST_EMAIL);
         var expectedPairwiseId =
@@ -150,16 +185,19 @@ public class CheckReAuthUserHandlerIntegrationTest extends ApiGatewayHandlerInte
     }
 
     @Test
-    void shouldReturn400WhenUserHasBeenBlockedForMaxPasswordRetries() {
+    void shouldReturn400WhenUserHasExceededMaxPasswordRetries() {
         userStore.signUp(TEST_EMAIL, "password-1", SUBJECT);
         registerClient("https://randomSectorIDuRI.COM");
 
-        redis.incrementPasswordCountReauthJourney(TEST_EMAIL);
-        redis.incrementPasswordCountReauthJourney(TEST_EMAIL);
-        redis.incrementPasswordCountReauthJourney(TEST_EMAIL);
-        redis.incrementPasswordCountReauthJourney(TEST_EMAIL);
-        redis.incrementPasswordCountReauthJourney(TEST_EMAIL);
-        redis.incrementPasswordCountReauthJourney(TEST_EMAIL);
+        var ttl = Instant.now().getEpochSecond() + 60L;
+        IntStream.range(0, 6)
+                .forEach(
+                        i ->
+                                authCodeExtension.createOrIncrementCount(
+                                        SUBJECT.getValue(),
+                                        ttl,
+                                        JourneyType.REAUTHENTICATION,
+                                        CountType.ENTER_PASSWORD));
 
         byte[] salt = userStore.addSalt(TEST_EMAIL);
         var expectedPairwiseId =
@@ -175,33 +213,7 @@ public class CheckReAuthUserHandlerIntegrationTest extends ApiGatewayHandlerInte
                         Map.of("principalId", expectedPairwiseId));
 
         assertThat(response, hasStatus(400));
-        assertThat(response, hasJsonBody(ErrorResponse.ERROR_1045));
-    }
-
-    @Test
-    void shouldReturn400WhenUserProfileNotFoundAndHasEnteredInvalidEmailTooManyTimes() {
-        userStore.signUp(TEST_EMAIL, "password-1", SUBJECT);
-        registerClient("https://" + INTERNAL_SECTOR_HOST);
-
-        var randomEmail = "random_email@email.com";
-
-        redis.incrementEmailCount(TEST_EMAIL);
-        redis.incrementEmailCount(TEST_EMAIL);
-        redis.incrementEmailCount(TEST_EMAIL);
-        redis.incrementEmailCount(TEST_EMAIL);
-        redis.incrementEmailCount(TEST_EMAIL);
-
-        var request = new CheckReauthUserRequest(randomEmail, "random-pairwise-id");
-
-        var response =
-                makeRequest(
-                        Optional.of(request),
-                        requestHeaders,
-                        Collections.emptyMap(),
-                        Collections.emptyMap(),
-                        Map.of());
-
-        assertThat(response, hasStatus(400));
+        assertThat(response, hasJsonBody(ErrorResponse.ERROR_1057));
     }
 
     private ClientSession createClientSession() {
